refactor(run-protocol): introduce `src/proposals/utils.js`
